Extract Attempt Diagnostics and Runtime Errors Types
