+++ Sources/ClawAgent/Runtime/AttemptRuntimePolicy.swift
@@ -6,21 +6,6 @@ package enum StreamingReattemptPolicy: Sendable, Equatable {
   case disabled
 }
 
-package struct ModelRoundTripObservation: Codable, Sendable, Equatable {
-  package let outboundModel: String
-  package let terminalModel: String?
-
-  package init(outboundModel: String, terminalModel: String?) {
-    self.outboundModel = outboundModel
-    self.terminalModel = terminalModel
-  }
-
-  enum CodingKeys: String, CodingKey {
-    case outboundModel = "outbound_model"
-    case terminalModel = "terminal_model"
-  }
-}
-
 package struct ProviderRoundTripAdmissionContext: Sendable, Equatable {
   package let roundTripIndex: Int
   package let priorRecordedTokens: Int
@@ -44,42 +29,6 @@ package struct ProviderRoundTripAdmissionContext: Sendable, Equatable {
 }
 
 package typealias ProviderRoundTripAdmission = BudgetDecision
-
-/// Payload-free attempt failure causes that a package caller may interpret under its own policy.
-/// This is descriptive only: it grants no retry or replacement eligibility.
-package enum AttemptFailureCause: CaseIterable, Sendable, Equatable {
-  case transportFailure
-  case credentialRefreshCompleted
-  case credentialRefreshExhausted
-  case credentialStateUnavailable
-  case deadline
-  case processInterruption
-  case partialStreamWithoutCompletedTerminal
-  case localOutputLimit
-  case modelIdentityMismatch
-}
-
-package struct AttemptDiagnostics: Sendable, Equatable {
-  package let outputCounts: AttemptOutputCounts?
-  package let modelObservations: [ModelRoundTripObservation]
-  package let failureCause: AttemptFailureCause?
-
-  package init(
-    outputCounts: AttemptOutputCounts?,
-    modelObservations: [ModelRoundTripObservation],
-    failureCause: AttemptFailureCause?
-  ) {
-    self.outputCounts = outputCounts
-    self.modelObservations = modelObservations
-    self.failureCause = failureCause
-  }
-
-  package static let empty = Self(
-    outputCounts: nil,
-    modelObservations: [],
-    failureCause: nil
-  )
-}
 
 /// Optional attempt-level controls used by package compositions and tests. The public runtime always
 /// receives `production`, keeping caller-specific policy out of the product API and gateway.
